Node & Electron work the same way, actually.
