+++ index.html
@@ -15,6 +15,8 @@
   <script src="node_modules/material-design-lite/dist/material.min.js"></script>
   <script src="node_modules/webcomponent-mdl/dist/webcomponent-mdl.min.js"></script>
   <script>
-    require('./renderer.js')
+    if (typeof(process.versions.electron) !== "undefined") {
+      require('./renderer.js')
+    }
   </script>
 </html>
