+++ main-node.js
@@ -1,0 +1,13 @@
+exports.main = () => {
+
+const path = require('path')
+const express = require('express')
+const port = 8080
+
+const app = express()
+
+app.use(express.static(path.join(__dirname, '.')))
+
+app.listen(port)
+
+}
