+++ main.js
@@ -1,34 +1,7 @@
-const electron = require('electron')
-const app = electron.app
-const BrowserWindow = electron.BrowserWindow
-const path = require('path')
-const url = require('url')
-
-let mainWindow
-
-function createWindow() {
-  mainWindow = new BrowserWindow({'width': 800, 'height': 600})
-  mainWindow.loadURL(url.format({
-    'pathname': path.join(__dirname, 'index.html'),
-    protocol: 'file:',
-    slashes: true
-  }))
-
-  mainWindow.on('closed', function() {
-    mainWindow = null
-  })
+if (typeof(process.versions.electron) !== "undefined") {
+  const main_electron = require('./main-electron.js')
+  main_electron.main()
+} else {
+  const main_node = require('./main-node.js')
+  main_node.main()
 }
-
-app.on('ready', createWindow)
-
-app.on('window-all-closed', function() {
-  if (process.platform !== 'darwin') {
-    app.quit()
-  }
-})
-
-app.on('activate', function() {
-  if (mainWindow === null) {
-    createWindow()
-  }
-})
