+++ package.json
@@ -5,7 +5,9 @@
   "main": "main.js",
   "scripts": {
     "start": "electron .",
-    "dev": "electron . --debug"
+    "dev": "electron . --debug",
+
+    "start-headless": "node ."
   },
   "repository": {
     "type": "git",
@@ -25,6 +27,8 @@
   "dependencies": {
     "grpc": "^1.1.0",
     "redux": "^3.6.0",
+
+    "express": "^4.14.1",
 
     "material-design-icons": "^3.0.1",
     "material-design-lite": "^1.3.0",
